+++ Example/android/app/src/main/java/com/goodeva/hub/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.goodeva.hub;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -13,7 +13,7 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "Example";
+    return "goodeva-hub";
   }
 
   /**
