+++ Example/android/app/src/main/java/com/goodeva/hub/MainApplication.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.goodeva.hub;
 
 import android.app.Application;
 import com.facebook.react.PackageList;
