+++ Example/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "Example",
-  "displayName": "Example"
+  "name": "goodeva-hub",
+  "displayName": "goodeva-hub"
 }
