+++ Example/ios/goodevahub/Images.xcassets/AppIcon.appiconset/Contents.json
