+++ Example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Example",
+  "name": "goodeva-hub",
   "version": "0.0.1",
   "private": true,
   "scripts": {
